Harden plugin web UI integration
