+++ crates/mesh-llm-host-runtime/src/protocol/mod.rs
@@ -11,6 +11,9 @@ pub(crate) use convert::*;
 use iroh::endpoint::Connection;
 use iroh::{Endpoint, EndpointAddr, EndpointId};
 use prost::Message;
+
+#[cfg(test)]
+mod config_tests;
 pub const ALPN_CONTROL_V1: &[u8] = b"mesh-llm-control/1";
 pub const ALPN_V1: &[u8] = b"mesh-llm/1";
 #[cfg(test)]
@@ -2010,76 +2013,6 @@ alias = "model-alias"
             "legacy-only and dual-encoded snapshots currently have distinct hashes"
         );
     }
-    #[test]
-    fn config_sync_full_config_roundtrip() {
-        use crate::plugin::{
-            GpuAssignment, GpuConfig, HardwareConfig, ModelConfigEntry, PluginConfigEntry,
-        };
-        let config = crate::plugin::MeshConfig {
-            version: Some(1),
-            gpu: GpuConfig {
-                assignment: GpuAssignment::Pinned,
-                parallel: None,
-            },
-            mesh_requirements: Default::default(),
-            owner_control: Default::default(),
-            telemetry: Default::default(),
-            defaults: None,
-            runtime: Default::default(),
-            models: vec![ModelConfigEntry {
-                model: "Qwen3-8B.gguf".to_string(),
-                mmproj: Some("mm.gguf".to_string()),
-                ctx_size: Some(8192),
-                gpu_id: None,
-                parallel: None,
-                cache_type_k: None,
-                cache_type_v: None,
-                batch: None,
-                ubatch: None,
-                flash_attention: None,
-                hardware: Some(HardwareConfig {
-                    device: Some("pci:0000:65:00.0".to_string()),
-                    ..Default::default()
-                }),
-                ..Default::default()
-            }],
-            plugins: vec![PluginConfigEntry {
-                name: "demo".to_string(),
-                enabled: Some(true),
-                web_ui_enabled: None,
-                command: Some("mesh-llm".to_string()),
-                args: vec!["--plugin".to_string()],
-                url: None,
-                settings: Default::default(),
-                startup: Default::default(),
-            }],
-            extra: Default::default(),
-        };
-        let snapshot = mesh_config_to_proto(&config);
-        let restored = proto_config_to_mesh(&snapshot);
-        assert_eq!(restored.version, config.version);
-        assert_eq!(restored.models.len(), 1);
-        assert_eq!(restored.models[0].model, "Qwen3-8B.gguf");
-        assert_eq!(restored.models[0].mmproj.as_deref(), Some("mm.gguf"));
-        assert_eq!(restored.models[0].ctx_size, Some(8192));
-        assert_eq!(
-            restored.models[0].gpu_id.as_deref(),
-            Some("pci:0000:65:00.0")
-        );
-        assert_eq!(
-            restored.models[0]
-                .hardware
-                .as_ref()
-                .and_then(|hardware| hardware.device.as_deref()),
-            Some("pci:0000:65:00.0")
-        );
-        assert_eq!(restored.plugins.len(), 1);
-        assert_eq!(restored.plugins[0].name, "demo");
-        assert_eq!(restored.plugins[0].enabled, Some(true));
-        assert_eq!(restored.plugins[0].command.as_deref(), Some("mesh-llm"));
-        assert_eq!(restored.plugins[0].args, vec!["--plugin"]);
-    }
-
     #[test]
     pub(crate) fn mesh_requirements_survive_owner_control_config_round_trip() {
         // Regression: NodeConfigSnapshot used to drop [mesh_requirements] on the
